+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - CI-419 - Commonised responsive behaviour for column-third
 
 ### Implemented enhancements
+- CMS-1774 - Add Invest region page, organise sass files
 - No ticket - Namespace ISD contact page styling to fix pollution on other pages
 - CMS-1734 - Added GA360 data to Perfect Fit views
 - CI-363 - Add anchor tags to all headings in setup guide articles